A way to do doubleEveryOther without the flag
In your class the other night, a couple of us mentioned that there was a way to do this with just pattern matching and no Boolean flag.

This is the way I wrote it. I use pattern matching to handle pairs of items at a time, which makes it easy to select which thing is doubled.

Since doubleEveryOther' is passed the reverse of the initial list, the first item in the list was originally the last, which is skipped, and the second item in the list is the next-to-last, which should be doubled.

So we just go down the list two at a time, until we hit either a singleton, or the empty list.

doubleEveryOther just reverses the result.

Since this is literally the first Haskell program I've ever written, I'm sure there are style issues in what I wrote. Feel free to correct these.